[level 0] Title: 옷가게 할인 받기, Time: 0.02 ms, Memory: 63.3 MB -BaekjoonHub

+++ 프로그래머스/0/120818. 옷가게 할인 받기/옷가게 할인 받기.java
@@ -7,6 +7,6 @@ class Solution {
         }else if(price >= 100000){
             price *= 0.95;
         }
-        return price;
+        return (int)price;
     }
 }
